pyproject.toml
[tool.poetry]
name = "tplinkmon"
version = "0.0.1"
description = ""
authors = ["Fergus In London <fergus@fergus.london>"]

[tool.poetry.dependencies]
python = "^3.9"
PyP100 = "^0.0.18"
prometheus-client = "^0.13.1"
click = "^8.0.4"
loguru = "^0.6.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
